// src/app/core/models/auth.model.ts
export interface LoginRequest {
  username: string;
  password: string;
}

export interface AuthResponse {
  access_token: string;
  refresh_token: string;
  expires_in: number;
  refresh_expires_in: number;
}

export interface User {
  username: string;
  accessToken: string;
  refreshToken: string;
}